WIP: Commander clean up.
